WeatherProvider: Improve provider initiation after boot
fix weather not loading sometimes.

We did set a 15sec interval but the api, by default, sets a check
interval for each location query to a longer time.
So with previous code, if the first check failed on the first attempt,
the api didn't try again even if the 15sec interval was still valid.
So we can use LocationRequest api to correctly trigger more checks in
our 20secs interval if the first one fails.
The setFastestInterval even allow to passively get the location
before the setInterval range if another app or the system try
to acquire location.

General code improvements: get all instances (APIs, builders, and so on)
only once instead of creating new ones on each query

Change-Id: I17d003fc91a34fcb4ec83e379a4d0138db89cc95
Signed-off-by: spezi77 <spezi7713@gmx.net>